Move tests into folders to match /src folder
